test(warnings): less fs dependency
Make sure the tests work on Windows and on the CI env where the checkout directory is not named
after the repo. Adding the package name to the plugin list, which won't be loaded as it can't be
found, as well as a relative path that can be found but won't be picked up as this package.