ci(actions): kill caching npm cache
The npm cache steps were added because Windows takes more than a minute to restore packages compared
to the ten seconds or so on the better OSes. With cache, Windows takes about 2 minutes and the other
OS are unchanged. This is ignoring the time taken to actually zip up and save the cache. Seems
pretty detrimental.