feat(orderwarning): ignore plugins included by their local paths
Newer versions of TS only allow plugins to be specified by package name. Dropped the tests and the
code for guessing when the plugin is included via a local path. This will make no difference to
anyone using the plugin normally via npm, and won't make any difference to properly set up
projects.